Refactor non-pipeline descriptor to be inside InfluxDbPublisher class
